server/requirements.txt
# Install with: `pip install -r requirements.txt`

Django==1.9.6
psycopg2==2.6.1
Pillow==2.9.0
gunicorn==19.3.0

# Facebook SDK
requests==2.9.1
facebook-sdk==1.0.0

# Rest Framework
djangorestframework==3.3.3
Markdown==2.6.6
django-filter==0.13.0

# CORS
django-cors-headers==1.1.0

# Google Cloud
gcloud==0.14.0
